Added MAP_TITLE and MAP_DESCRIPTION as CustomProperties to our maps. Right now, these will be rendered at the start of each map in the middle of the screen. Might change that later on.
Introduced the NARRATOR as an IUpdateable, IRenderable that's contained in an environment. Might even move this to Game level. It functions quite similar to the SpeechBubbles and displays text in a static box independently from the RenderScale. Lots of setters allow you to adjust colors, font, or even background images for the text box.

git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@2373 16493bb7-3577-b145-b9c3-72a97d279fb6