Re-introduce cache for static tiles during map rendering process
When rendering a tile map, it is only possible to evaluate animated tiles every frame because they might change.
Since static tile aren't supposed to be changed during runtime, they can be cached in order to increase the rendering performance.
Nevertheless I've introduced a new graphics configuration which allows to disable this caching if necessary.
Renamed some render methods of the IMapRenderer interface to better represent what they are doing.
Implemented a helper method that allows to check whether a specific tile on a map does have an animation.

Fixes #6