The resolution now scales the game depending on the available or configured resolution. In fullscreen, the specified resolution is ignored and the game is rendered at the max available size of the current display. Implemented the possibility to create a custom Resolution for e.g. fullscreen.
Typically, the games made with the litiengine do have some kind of rendering scale because plain pixel art is very small on a full-HD display. Also the rendering onto the buffer-strategy prevents the possibility of scaling up the graphics object (remember: no openGL is used).

Nonetheless this optimizes the rendering process specifically for pixel art kind of games.