Implement support for RenderType on particle level
Extend Particle to support a ParticleRenderType that defines whether it should be rendered with the rendertype of the emitter, as groundRenderable or overlayRenderable.
Emitters now provide two IRenderable implementations that are responsible for rendering all particles that have their ParticleRenderType set to GROUND or OVERLAY.
Extend the environment to automatically add/remove the provided IRenderable instances when a emitter is added to or removed from the envionment.