No need to explicitly exclude dead entities from collision
By default collision will be set to false anyway when an entity dies. By not enforcing this condition anymore, it is now possible to have e.g. props that keep collision, even when they're destroyed.
Fix an issue with prop's health not being set for the actual health value but only the max health value.