Fix issue with the sound playback not being properly closed
Cancel for the playback run method was not set.
Refactored the SoundSource implementation to implement Runnable itself instead of having the PlayRunnable implementation.
This refactoring was based on the fact that the runnable implementation used a ton of fields from the SoundSource itself.

Improved the implementation of the SoundDataLineCloseQueue by extacting the closeAllSoundSources method.