Major overhaul on the resource management
It is no longer necessary to have a sprites.info file; It is still possible though to import sprites from such a file.
Spritesheets can now just be imported to the editor and will be serialized in the game file.
The spritesheet files are no longer kept in the game file. You need to re-import the file in case, you've been using it.

Extend the editor with a view that displays the assets.
Extended the icon tree implementation so that it can be reused in the asset view.

Most of this is still WIP and will be improved in the next few days.

Extend the ImageCache with a clearAll method.