Clean up static shadow implementation
It is now possible to set the offset for a static shadow.
It is now possible to set the color for static shadows on the map.
Clean up LightSource implementation by removing redundant brightness, which was just the alpha value.
Extend the CustomPropertyProvider to allow specifying default values for retrieving string properties.