Clean up IRenderEngine interface
All methods that depend on the current camera (render the respective object relative to the viewport) are now provided directly on the IRenderEngine interface instead of being static methods on the RenderEngine class.
This way, its easier to determine which methods take the camera into account internally and which don't.