Update the creature animation controller to not use the entity name as sprite prefix.
Major refactoring and clean up on the way, the CreatureAnimationController works.
It is now possible to define a spritePrefix for a creature which will be used to determine the spritesheets for the animation controller.
Remove the DecorMobAnimationController because it basically just did the same thing as the CreatureAnimationController.

Implement the possibility to provide the sprite prefix for creatures as Annotation.
As a fallback the Creature implementation takes the class name.