Implement the next major steps to directly load creatures from map objects.
Extend the IImageEffect interface with a priority that determines in which order effects get applied.
Provide a basic CreatureMapObjectLoader.
Start to implement the CreaturePanel in the editor.

Refine the entity rendering process to take the actual sprite size into account -> image effects might resize the sprite.
Implement the possibility to remove image effects from an AnimationController.
Improve the borderAlpha implementation by actually only returning the border.

Make use of the LabelListCellRenderer for the DecorMobPanel.