Provide some additional API sugar
Extend the IScreenManager with the possibility to get its center Point.
Extend IMouse with method to set location with two double values.
Properly handle the potential exception that can occur when the Robot for the Mouse cannot be initialized. Since the Mouse component will be unusable in this case, its okay to handle this with an exception.