Less boilerplate for loading map objects
This commit reduces the amount of work that is necessary to load custom entity implementations by a specific mapobject type.
Provided a reflection based CustomMapObjectLoader that supports constructors with IEnvironment and IMapObject parameters, just an IMapObject parameter or an empty one.
Extended the Environment with a method that allows to register custom entity types to support being loaded from a map.
Rename the addMapObject method to "load" and made it public.
Rename ICustomPropertyProvider.getAllCustomProperties to "getCustomProperties".
Added documentation for the newly introduced methods.
Added unit tests to ensure that loading loading custom entities works properly.

Issue #127