Refactor entity controller logic
The Game no longer holds a static EntityControllerManager.
Instead, every Entity gets a new EntityControllers object that manages the controllers for the instance.

The implementation has been done in such a generic way that it can be used similar to a component system.
There are three known default types of controllers in the IEntityAnimationController, IBehaviorController and IMovementController.
The default IAnimationController interface is a bit more general and doesn't require any entity. Only controllers that are related to an entity are managed this way.

Developers can reuse the new logic to define instances of the known types, define custom ones or overwrite preset instances.

The environment will automatically call the controllers attach/detach methods as entities get loaded to or unloaded from the environment.