Re-work the rendering pipeline to make it more flexible
With this commit, the whole pipeline that defines how everything is rendered on the Environment has been improved.
Basically the whole rendering process is partitioned by the RenderType enum values.
The registering process on the Environment has been unified: IRenderable implementations can now be registered for any of these enum values.
BACKGROUND -> GROUND -> SURFACE -> NORMAL -> OVERLAY -> UI

The render events have been reworked and now depend on the RenderType.
The IMapRenderer has been reworked and now depends on the specified RenderTypes. It is possible to render any layers of a particular or multiple types.

Improved the implementation of Map and Layer by not creating a new ArrayList every time the getters for the Layers were accessed.

Closes #128