Clean up the RenderComponent implementation.
It is designed to always render the current screen of the ScreenManager. Therefore, there is no need to provide the current screen externally as IRenderable.
Also the RenderLoop provided the potential to register multiple IRenderable instances which is by design not intended. Only the currently active Screen should be rendered. All further rendering can be done by implementing a custom Screen implementation and overwriting the render(Graphics2D) method.

Remove the applyRenderScale field from the ScreenInfo annotation because the scale is handled by the environment and no the screens anymore.