Implement MapObjectSerializer
This implementation uses the TmxProperty annotation and reflection to determine which fields of the IEntity implementation need to be serialized to a custom property of the MapObject.
Replaced the CustomMapObjectProperty annotation by the TmxProperty annotation. This new annotation is defined on the actual field which is more practical.
Also this annotation directly links the field to a key value for custom properties. Previously this was a very loose connection and one had to know the related field.

Annotated all currently implemented custom properties for the engine's default entity types with the TmxProperty annotation.
This will also allow to further automate the MapObjectLoaders in the long run.

Implement a unit test that ensures that the MapObjectSerializer works properly.
Rename some MapObjectProperty constants to better reflect their use case.

Closes #66