Improve accessibility
First, there’s no reason for this class to be final.
Second, the constructor should be public, so that it can be instantiated by other parts of the application, e.g. to place a wrapper around it.
Third, make ConstructorInvocation protected rather than private so that subclasses can define their own invocations.