Extend the CustomPropertyProvider implementation
It now provides new overloads that allow to set values for primitive types (and enum values) without having to convert them to a string first.
Also simplified the method naming according to the changes done to the getter methods.

The general Entity implementation now has the possibility to provide custom properties. Previously this was only implemented for the Trigger class.