Implement support for infinite maps.
Extend XML model to be able to parse both the usual maps but also the infinite maps that are based on chunks.
With this commit, the unmarshalling already works. The LITIengine itself is not treating the infinite maps any differently which means that it internally just creates a map that wraps all the chunks.
TODO: Fill the areas where no chunks are defined with empty tiles.

Add a basic unit test to ensure that the unmarshalling works in general.