Implement support for infinite maps
- Extend layers with a backreference to their map
- Extend mapobjects with a backreference to their layer
- Since the litiengine handles infinite maps a bit different than the tiled editor (the map in the engine always has an explicit boundary and the tiles start from 0,0), the mapobjects need to account for this when presenting their coordinates.
  The mapobject coordinates are serialized in the original tiled manner and the engine incorporates the chunk offset when dealing with infinite maps.

This is not the prettiest solution but dropping the assumption that a map has a boundary would imply rewriting very large parts of the engine which is why I decided against it. In the LITIengine, the tiles of a map always start at 0,0  while infinite maps in the tiled editor might provide tile locations "outside of the maps bounds" (i.e. there are no bounds to the map anymore which is completely contradictory to the way the LITIengine works).
For this reason, the LITIengine translates the Tiled Infinite Maps to a large map that wraps all the chunks.

Issue #150