Fix warnings UtiLITI
Chances are, these classes will never be serialized, so we can safely
suppress the warnings.
For FileDrop, there's no way of avoiding the unchecked cast, and it's
guaranteed to be a List<File>, so just suppress it.
The raw type warnings can be fixed by simply using wildcards, so they
aren't suppressed.