Improve the Game API
Provide further documentation on the Game API.
Improve the GameMetrics implementation.
Extend the GameWindow with getWidth() and getHeight() methods instead of getResolution().getWidth() and getResolution().getHeight() which allows a more direct access for a very common use case of the GameWindow.
Make use of these new methods.
Rename getScreenLocation() to getWindowLocation().