Make it possible to retrieve resources via an alias.
This allows to directly access a map by its name instead of having to identify it via it's full resource path.
Right now this is only implemented for maps because we require map names to be unique for a LITIengine game anyway.

Remove the identifier from the IEnvironment. It was previously added for this branch but has now been replaced by a mapName based logic.
Add documentation to the Game.world() method.