Drop IPhysicsEngine interface.
Since the interface was an exact representation of what the implementation provided and there are no plans on making this part of the engine interchangeable, there's no point in keeping the additional interface.
Also reorganize the methods of the PhysicsEngine.

The provided API stays exactly the same so there are no breaking changes here for existing games.