Improve the GameMetrics
With this change rendering the game metrics has been improved and additional information have been added.
Added information about the max. estimated fps, the java version, the currently active threads.
Previously the "logDetailedRenderTimes" option spammed a lot of information to the console about the rendering performance of the environment. If the new option "trackRenderTimes" is now enabled, the GameMetrics will now output this diagnostics information right on the screen, just like the other metrics.

Fixed a potential exception in the environment when removing entities.