Implement support for TextureAtlas.
With this change, it's possible to import a TextureAtlas created with the TexturePacker tool (https://www.codeandweb.com/texturepacker).
The LITIengine supports the generic XML format (without trimming/offsetting). The algorithm has to be either "Basic" or "MaxRects".

Added unit tests to ensure that everything works correctly.

Improve the image rotation algorithm. It now keeps the image dimensions.

Fixed a few potential exceptions of the utiLITI editor.
Implemented possibility to import sprites in the utiLITI editor via TextureAtlas.