Generalize the handling of custom Entity actions.
With this change, it's possible to register custom action, that can e.g. be used by EntityController implementations to execute further behavior without having to know too much details about the actual entity implementation.
It's basically a Map that identifies callbacks by a certain name.
The first usage of this would be the "jump" mechanism that is now provided by default in the PlatformingMovementController.
If an entity implementation provides a "jump" method with an Action annotation, this method will be automatically registered by the Entity implementation and can thereby be used later via Entity.perform("jump") (which is exactly what the PlatformingMovementController does.

The Entity implementation uses reflection when the constructor is called to evaluate what action are available on the class. Then it registers these actions.
It's also possible to explicitly register custom action by calling Entity.register(String name, Runnable action). This will overwrite any previously registered action for the specified name.