Add OpenSans as global editor font.
Drop the additional border for the panels, since this information is redundant, now that we have tab-pages.
Start to specify helper methods to setup the listeners on the MapObjectProperty Controls.
Make it possible to directly get a font in a certain size or style from the resources.
Rework the layouting of the property panels (started with CollisionPanel).