Drop the IEnvironment interface.
It didn't have any real purpose apart from initially designing the API of the Environment class.
Make the Environment class final. In previous versions it was designed to be overwritten which is no longer the case. Instead it's now one of the infrastructure classes that the engine provides and it's now designed to be flexible and extendable instead (e.g. via registering additional MapObjectLoaders.