Replace "isObstacle" flag by CollisionType.
Drop special "isObstacle" handling in several places which required some casting and custom handling.
Adjusted the UI of the utiLITI editor to reflect the changes.

Fix default loading of MapObject properties: The default loading mechanism now also sets the fields of parent classes if properly annotated with TmxProperty (e.g. the collision properties are now automatically set for a Creature).
This now allow for a simpler implementation for all the MapObjectLoaders.