Change datatype for entity width/height and angle.
Previously the datatype was float but the datatype for x and y were double. So it was essentially a mix of concepts with the downside to explicitly provide float syntax when accessing these properties. This unifies the API and simplifies the usage while the performance benefit is small enough not to matter for the engine.

It's now no longer possible to manually add a Rectangle2D to the PhysicsEngine. This was an old approach to "static" collision which can now be achieved by using a CollisionBox instead
Extend the CollisionBox with more convenient constructors.
Add the collisionType to the CollisionInfo annotation.