Clean up some code regarding grid settings in the editor
Setting the grid width and height explicitly didn't render a different grid because the grid size is strictly related to the shape of the tiles (especially in non orthogonal map orientations). The custom grid size was only applied for snapping (which btw. is also a bit weird in other orientations). So essentially it snapped to an invisible grid if the width and height were any different from the tile dimensions which makes no real sense.

Remove explicit private variables and replace them by the usage of the configuration.