Add some additional methods to retrieve entities by layer.
Fix a bug where the serialization process assumed that a MapObject always has to have a MapObjectLayer as parent.
Add method overload to load a game resource file from a String to stay backward compatible.
Make use of "getEnumValue" to retrieve the light shape.
Extend the getEnumValue process to be case insensitive as a fallback.