Remove the toolbar
The buttons on the toolbar are not really required and can be accessed elsewhere in a more convenient way.
Reworked the "MOVE" edit-mode which can now be triggered by clicking the mouse button within the bound of a map object instead of having to press CTRL or manually activate it.
Fix an issue that caused the transform controls not to be rendered properly upon undoing a move command.
Rework some of the menus.
Add the "Add" Submenu to the Edit menu.