Clean up "render entity with layer" feature
With this change, it's now an explicit property that can be set on a MapObject/Entity. Since rendering the entity with it's layer is a completely different approach to rendering it with our predefined RenderTypes it's now necessary to explicitly opt-in to this behavior instead of the engine assuming to render an entity with the layer if their render type matches.
Set the default RenderType for Layers to "NORMAL" since this is considered the default value in the engine.
Prepare the MapObjectLoader to support the new custom properties RENDERWITHLAYER and RENDERTYPE.