Default render type should be GROUND for ALL layers.
If different types of layers default to different render types, it
screws with the layering for layers without set render types. Yes, it
means that layer-bound entities will have a different default render
type, but there's no perfect way to handle this, and it's more important
that the original layering of the map be preserved.