Revert "Fix image layer rendering"
This reverts commit bf90ac9c6b668ac8b257aa50cb62a90c7c5dce5b.

While the general idea of this change is reasonable, it (at least) broke the image layer rendering for maps from a ResourceBundle. ImageLayers could no longer be rendered in the utiLITI editor and games that use a ResourceBundle.
The images that a tileset/map referes to are typically relative to the map and not the application directory. e.g. if I have an image layer that refers to "image.png" and the maps are located under a "maps" directory, the changes don't account for any sub-folders because they don't acknowledge the fact that the images are not relative to the application directory. Also, using Spritesheets allows for the resources to only be present in the ResourceBundle. After the change, this was no longer possible.