Fix map serialization
Revert switch to URLAdapter for Tileset because this broke the tmx (XML) compatibility of our serialization which also resulted in paths from the local machine being serialized to the resource bundle or when exporting a map.
Implement unit tests that ensure the output of our tmx serialization is equal to the input.

If we want to switch to URLs here at some point we need to make sure that we don't break the litiengine's tmx format.

Closes #284