Consolidate the RenderLoop with the GameLoop
The concept of separated loops for gamelogic and rendering is not really of much use for the kind of games that are/can be created with the litiengine. It causes a lot of concurrency issues that can be simply resolved by performing the game logic and rendering sequentially. e.g. Position updates in the game loop could cause visual stuttering in movement and positions of entities could thereby change inbetween a frame which caused inconsistent behavior.

Moreover this simplifies the general concept of loops for the LITIengine and the developer doesn't always have to ask himself which loop should be used for a task.
Overall, its a more straightforward and less error-prone approach which should also be easier to understand.