Implement entity render events for individual entities.
Extend the entity with the possibility to register render listeners directly.
Forward the explicitly called events from the RenderEngine to the listeners on the entity.
Try to keep the IEntity interface as clean as possible in this process by only implementing the EntityRenderListener interface in the Entity class instead of extending it on the IEntity interface.
This implementation is only used internally to forward the events of the RenderEngine and should therefor not be used explicitly by the developer.

Cleanup some listener naming and implementations.
Closes #301