Remove MapObjectException
In the LITIengine, we don't want to use exceptions as a mean of "normal" control flow. Avoiding them if the case can somehow be handled is a basic concept of the engine to ensure that a game doesn't unnecessarily crash.
In this particular case, the Environment implementation just catched the exception anyway.

The changed implementation also provides some more details on how CustomMapObject types can be registered and what conditions need to be met.