Account for resolution scaling at runtime
When the window size gets changed at runtime, the resolution scale was previously not updated. Also we just adjusted the base render scale of the game which is technically not correct.
With this change and with the option "gfx_enableResolutionScale" set, it's now possible to dynamically adjust the game's render scale based on the window size.