Update camer use in Render&sound engines
Render now use ActiveCamera, assuming all use of these functions in Render are for rendering objects on the game map, during the game world render phase.

When it comes to sound, it is ambiguous how sound should be heard when there are more than one active camera, so for now it is left as is, using the main camera for translating the audio source.