Implement support to explicitly set tiles
Previously, it was necessary to provide static TileData that was not considered to be changed afterwards. With this change, it's possible to change tiles on a map at runtime.
The main API for this is ITileLayer.setTile.

Don't throw InvalidTileLayerException, try to avoid crashing a game over all other code-design.

Issue #139