Add API that supports generating maps from code
This API is designed to ensure that the "finish" logic of a map is always called when generating maps from code. It also simplifies the generation of of tile layers with a method that provides a simple callback for (x, y) coordinates to decide which tile should be added at a particular grid location.