Fixed the handling of the anim controller affine transform
With this change, the RenderEngine will take care of the offset in which the image is being rendered for an entity instead of incorporating it into the AffineTransfrom instance.
This way the AffineTransform instance of an AnimationController only needs to account for the actual changes it wants to carry out and will then be concatenated with the Camera translation transform during the rendering process.
This includes any offset due to divergent dimensions between an entity and the respective image.

PR #324