Set the collision box size for Triggers implicitly
The CollisionBox of a Trigger is always intended to be equal to its BoundingBox. However, when creating a trigger via code instead of via MapObjectLoader (from the map directly) this was previously not guaranteed and it required the size to be set explicitly in order for the Trigger to be usable.