Fix bug with mousewheel scrolling
This issue caused unexpected behavior when scrolling via mouse wheel and locked the camera to the boundaries of the map without the possibility to scroll to a different place via mousewheel. This was mainly caused due to a missing flag which caused the implementation to reset the focus multiple times during an update operation.
With this change, the scrolling logic has also been moved to a dedicated class with an abstraction layer to the actual Swing components. This allows us to write tests for this part of the editor and makes the code better readable.