Improve the performance of the GridRenderer.
Previously this renderer just rendered the Grid by rendering each tile as Shape on each frame. Shape rendering in Java is for some reason very costly compared to rendering an image and the grid was often the most expensive rendering operation in the editor.
With this change, the GridRenderer caches matching grid images for different zoom levels and just renders these images.