Replace Gradle Natives plugin with a manual approach
In an effort to get the jinput native binaries extracted, I decided to get rid of the external Gradle Natives plugin which has been abandoned in the meantime.
Unfortunately the custom "natives-all" classifier doesn't seem to work in correlation with settings.gradle.kts and the VersionCatalogBuilder.
Seems like classifiers are unsupported - at least I didn't find any way to achieve it which is why had to manually add it to the build.gradle.kts file.