Pass through event registrations to the current gamepad
Fix issue where registering and unregistering an event on the current gamepad didn't work on-the-fly.
The listeners were only hooked up once as a whole but adding/removing individual events at runtime was not supported properly which also made it a matter of execution order because events would have to be hooked up before the first Gamepad was detected.