Handle invalid color values more gracefully
Overall the handling of color values in utiLITI in particular was very fragile and even caused the editor to crash if an unexpected value was present in a project or map file.
This change handles these cases more gracefully and keeps the editor running while still providing a proper error that notices the wrong color string format.